pyproject.toml
[project]
name = "hand-age"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "openai>=1.107.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
